repos/LogicNet-Subnet__LogicNet/requirements.txt
bittensor==9.0.0
bittensor-cli==9.0.0
Pillow==10.2.0
PyYAML==6.0.1
setuptools==70.0.0
slowapi==0.1.8
tqdm==4.65.0
httpx==0.26.0
numpy==2.0.1
openai==1.35.14
sentence-transformers==3.0.1
python-dotenv==1.0.1
loguru
sympy
wandb
datasets
minio